move beerdescription and watermeasurement to misc package
