+++ src/main/java/nwoolcan/model/brewery/production/batch/ModifiableBatchInfoImpl.java
@@ -1,6 +1,8 @@
 package nwoolcan.model.brewery.production.batch;
 
 import javafx.util.Pair;
+import nwoolcan.model.brewery.production.batch.misc.BeerDescription;
+import nwoolcan.model.brewery.production.batch.misc.WaterMeasurement;
 import nwoolcan.model.brewery.production.batch.step.parameter.Parameter;
 import nwoolcan.model.brewery.production.batch.step.parameter.ParameterTypeEnum;
 import nwoolcan.model.brewery.warehouse.article.IngredientArticle;
