+++ src/main/java/nwoolcan/model/brewery/production/batch/misc/BeerDescriptionImpl.java
@@ -1,4 +1,4 @@
-package nwoolcan.model.brewery.production.batch;
+package nwoolcan.model.brewery.production.batch.misc;
 
 import javax.annotation.Nullable;
 import java.util.Objects;
