+++ src/main/java/nwoolcan/model/brewery/production/batch/misc/WaterMeasurement.java
@@ -1,4 +1,4 @@
-package nwoolcan.model.brewery.production.batch;
+package nwoolcan.model.brewery.production.batch.misc;
 
 import nwoolcan.model.brewery.production.batch.step.parameter.Parameter;
